Fixes for non-renamed tags
