+++ buildinfo/vaadin/avatar-group/test/avatar-group.test.js.patch
@@ -101,6 +101,15 @@ Index: test/avatar-group.test.js
        expect(items.length).to.equal(2);
      });
  
+@@ -144,7 +145,7 @@
+       group.items = group.items.slice(0, 2);
+       group.maxItemsVisible = 1;
+       await nextRender(group);
+-      const items = group.querySelectorAll('vaadin-avatar:not([hidden])');
++      const items = group.querySelectorAll('vaadin24-avatar:not([hidden])');
+       expect(items.length).to.equal(2);
+     });
+ 
 @@ -177,7 +178,7 @@
        group.items = [];
        group.items = items;
@@ -128,6 +137,15 @@ Index: test/avatar-group.test.js
        expect(items.length).to.equal(3);
        expect(overflow.abbr).to.equal('+3');
      });
+@@ -227,7 +228,7 @@
+       group.set('items', group.items.slice(0, 2));
+       group.style.width = '50px';
+       await onceResized(group);
+-      const items = group.querySelectorAll('vaadin-avatar:not([hidden])');
++      const items = group.querySelectorAll('vaadin24-avatar:not([hidden])');
+       expect(items.length).to.equal(2);
+     });
+ 
 @@ -251,7 +252,7 @@
        overflow.click();
        await oneEvent(overlay, 'vaadin-overlay-open');
