+++ buildinfo/vaadin/date-time-picker/test/validation.test.js.patch
@@ -20,7 +20,7 @@ Index: test/validation.test.js
  }
  
 -customElements.define('vaadin-date-time-picker-2020', DateTimePicker2020Element);
-+internalCustomElements.define('vaadin-date-time-picker-2020', DateTimePicker2020Element);
++internalCustomElements.define('vaadin24-date-time-picker-2020', DateTimePicker2020Element);
  
  const fixtures = {
 -  default: '<vaadin-date-time-picker></vaadin-date-time-picker>',
