+++ buildinfo/vaadin/grid/test/column-auto-width.common.js.patch
@@ -44,6 +44,15 @@ Index: test/column-auto-width.common.js
    });
  
    it('should have correct column widths when the grid is visually scaled', async () => {
+@@ -200,7 +201,7 @@
+ 
+     fixtureSync(`
+       <style>
+-        vaadin-grid::part(cell) {
++        vaadin24-grid::part(cell) {
+           padding-right: 20px;
+         }
+       </style>
 @@ -228,9 +229,9 @@
  
    beforeEach(() => {
