+++ package.json
@@ -5,10 +5,11 @@
   "private": true,
   "scripts": {
     "start": "web-dev-server --node-resolve --open /dev",
-    "build": "npm run build-element-meta && npm run build-packages",
-    "build-packages": "node scripts/build.js",
-    "build-demos": "node scripts/build-demos.js",
-    "build-element-meta": "node scripts/build-element-meta.js",
+    "build:clean": "rimraf ./git_modules && rimraf ./packages && rimraf ./buildinfo && npm run clone && npm run build",
+    "build": "npm run build:element-meta && npm run build:packages",
+    "build:packages": "node scripts/build.js",
+    "build:demos": "node scripts/build-demos.js",
+    "build:element-meta": "node scripts/build-element-meta.js",
     "test": "web-test-runner",
     "test:watch": "web-test-runner --watch",
     "test:chrome": "web-test-runner --config web-test-runner-chrome.config.cjs",
