--- scripts/supplemental-element-names.js
@@ -1,6 +1,0 @@
-/**
- * additional string literals to treat like tag names
- */
-export const supplementalElementNames = [
-  "vaadin-${type}-picker", // vaadin-date-time-picker creates a tag dynamically
-];
